improved exceptions
